Skeleton & Basic Algorithm for Sniffing Numbers
I created this file to write down the bare-bones method to sniff numbers
in a string. This program uses string slicing, though it can be easily
modified to use other methods too.

In sniffing numbers out of strings, the basic idea is that numbers are
strings of digits in sequence.
So, if we find a number, we check if the subsequent digits are numbers
too. When we find a character that is not a number, we stop. In this way
we get a sequence of digits.. which may then be converted into a number
using float/int on a string slice, or float/int on each character &
adding it to the already formed number (num*10+float(a[i]).