Fixed a problem with sniffnums4
In sniffnums4 if one entered a number with a sign right after it, it was
not counted as a number, because float cannot convert number-sign into a
number.
So, instead, I changed the code allowing signs only initially. If the
try for the first character succeeds, and there is no sign, any
subsequent occurence of a sign in a number breaks the search, as it
should.