Fixed some code & wrote sniffnums5
sniffnums5 doesn't recognize signs but otherwise it does a good job
sniffing numbers out of strings. Plus i made some overall changes to the
code, specifically in the section where we find a number and print the
first line:: "The numbers in the string are:". Now, all functions print
the same lines to the screen & use nearly the same variables for the
same purposes.