Fixed sniffnums5 to also accept signs.
Now with this file we have a set of functions that can sniff numbers out
of files, in diff ways.
I'll also save a copy of this file as number_in_strings.py. If I talk
about a file with functions to deal with numbers in strings it will be
this.