numbers_in_strings.py
This is all the code that I've written to sniff numbers out of strings.
Note: All the code in the functions of this file are actually
templates:: They sniff out a number, and then print it. You can replace
that printing statement with any code that can process the sniffed
number. These functions print & count the numbers. You might save them
in an array, or do anything with them. For that, put your code after the
part n=float(string) or when the comment says, "we have a number n..."
I plan on writing a file with the algorithm for sniffing numbers with
signs. There is also a skeleton file for the basic code. I'll put the
algorithm in that file itself.