Added the basic algorithm to generate primes.
A number is prime if it is not divisible by any number other than 1
&
itself. Plus, the largest factor of a number (other than itself)
is
lesser than or equal to half the number.  These combined together
form
the basis of the prime finding algorithm.
Plus, Since I learned
lists last night, I wrote a program to find primes
using lists. That is
here too. In this, the program generates a basic
list of primes using
our earlier algorithm, and then when more primes
are needed it uses the
list. Every non-prime number can be expressed as
a product of primes
that are lesser than it. Plus, the fact that the
largest factor of a
number has to be less than half the number itself.
It finds the first
10k primes in ~20s. The first 100K in ~30 min. The
code isn't final and
will change & improve(maybe).