100,000 in 38 Secs
This new function uses the p>n/p check to stop looking for factors,
besides already skipping ones that have factors. It finds the 1
millionth prime in 20min. THe 1 millionth prime is 15485863.
I've noticed that if you remove all assignment operations that store the
generated primes, or any print statements that print the generated
primes the speed improves. The times stated are those without the a=i
statements.(i.e we use return i instead, since the loop terminates as
soon as c=n. Moreover the method also finds the sum of all primes under
2 million in 1:20s (Project Euler #10).