Using p^2>n in prime factorisation
I made two improvements to the process of prime factorisation in
primefactor_v2. #1 reduces the number of check-cycles starting with 1.
Here we sweep once from the number 1 toward n, and factor out the
factors as we find them. #2 Stops the check process when we have reached
a number p such that p>n/p which tells us that n is prime. This factors
out the product of two 10 digit primes in 1:30 Hours.