Tried Implementing the Sieve Approach
Starting with a list of numbers from 2 till n, we remove every thing div
by 2, then all div by 3 & so on. For each number in the list, we don't
remove the number, but every other number that it divides. Since we have
consecutive numbers, numbers divisible by p, are spaced in intervals of
p. Using this idea, we end up with a list of primes spaced out by
zeroes. For the first 1 million primes this takes 29 secs to calculate.
However, printing & displaying the results takes time. Moreover, we
cannot stop once we have generated a specific number of prime numbers.
Plus, the size of the list, which we define first is fixed. One can use
the del method with a while loop to reduce the size of the array too.
while(i<len(a)):
if(a[i]==0):
del a[i]
else:
i=i+1
The len(a) varies as elements are deleted from the loop. Or, you could
adjust all the primes in order to one end of the loop, and slice the
other end.