A Few Hackerrank problems.
Sat down to solving a few of the basic problems on the warmup list. Most
of this code was typed into the browser window & got accepted on the
first run. What I dont understand is why I sometimes get the first rank,
and other times near 240. Anyway, most of the code I wrote needs review.
I used lists to solve the problems. THese problems pre-tell a N value so
that an array can be declared, but Python allows one to create a  one
dimensional array: a list, and append objects to it. Anyway, will have
to submit most problems again when I learn a second language.